CASMCMS-9225: Move HSM client to new paradigm
