+++ src/bos/common/clients/hsm/exceptions.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 #
 # MIT License
 #
@@ -22,40 +21,13 @@
 # ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 # OTHER DEALINGS IN THE SOFTWARE.
 #
-import logging
-
-from bos.common.values import Action, Status
-from bos.operators.base import BaseOperator, main
-
-LOGGER = logging.getLogger(__name__)
 
 
-class GracefulPowerOffOperator(BaseOperator):
+class HWStateManagerException(Exception):
     """
-    - Enabled in the BOS database and the status is power_off_pending
-    - Enabled in HSM
+    An error unique to interacting with the HWStateManager service;
+    should the service be unable to fulfill a given request (timeout,
+    no components, service 503s, etc.); this exception is raised. It is
+    intended to be further subclassed for more specific kinds of errors
+    in the future should they arise.
     """
-
-    retry_attempt_field = "power_off_graceful_attempts"
-
-    @property
-    def name(self):
-        return Action.power_off_gracefully
-
-    # Filters
-    @property
-    def filters(self):
-        return [
-            self.BOSQuery(enabled=True, status=Status.power_off_pending),
-            self.HSMState(),
-        ]
-
-    def _act(self, components):
-        if components:
-            component_ids = [component['id'] for component in components]
-            self.client.pcs.transitions.soft_off(component_ids)
-        return components
-
-
-if __name__ == '__main__':
-    main(GracefulPowerOffOperator)
